% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_tiles.R
\name{get_tiles}
\alias{get_tiles}
\title{Retrieve tinitaly DEM tile names that intersect with sp object}
\usage{
get_tiles(x)
}
\arguments{
\item{x}{sp object as defined in package sp}
}
\value{
character vector of tinitaly tile names that intersect with supplied sp object
}
\description{
Retrieve tinitaly DEM tile names that intersect with sp object
}
\examples{

sp <- sp::SpatialPoints(cbind(1038553, 4302213))

tiles <- get_tiles(x = sp)

}
\references{
Tarquini S., Isola I., Favalli M., Battistini A. (2007) TINITALY, a digital elevation model of Italy with a 10 m-cell size (Version 1.0) [Data set]. Istituto Nazionale di Geofisica e Vulcanologia (INGV). \url{https://doi.org/10.13127/TINITALY/1.0.?\200\235}
}
